move ok api response to the correct file
